refactor(storage): split db.go into focused module files
