+++ internal/storage/db_merge.go
@@ -1,0 +1,115 @@
+package storage
+
+import (
+	"context"
+	"database/sql"
+	"time"
+
+	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/model"
+)
+
+// MergeWindow represents a group of consecutive recordings eligible for merging.
+type MergeWindow struct {
+	StartTime    time.Time `json:"start_time"`
+	EndTime      time.Time `json:"end_time"`
+	SegmentCount int       `json:"segment_count"`
+	Format       string    `json:"format"`
+}
+
+// MergeAndReplaceRecordings atomically inserts a merged recording and deletes old recordings in a single transaction.
+// This reduces SQLITE_BUSY contention compared to separate INSERT + SetMerged + DeleteBatch calls.
+func (d *DB) MergeAndReplaceRecordings(ctx context.Context, merged *model.Recording, oldIDs []string) error {
+	if len(oldIDs) == 0 {
+		return d.InsertRecording(ctx, merged)
+	}
+	tx, err := d.db.BeginTx(ctx, nil)
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+
+	q := `INSERT INTO recordings(id, camera_id, file_path, format, started_at, ended_at, duration, file_size, frame_count, merged) VALUES(?,?,?,?,?,?,?,?,?,?);`
+	_, err = tx.ExecContext(ctx, q, merged.ID, merged.CameraID, merged.FilePath, merged.Format, timeToDB(merged.StartedAt), timeToDB(merged.EndedAt), merged.Duration, merged.FileSize, merged.FrameCount, true)
+	if err != nil {
+		return err
+	}
+
+	for _, id := range oldIDs {
+		_, err = tx.ExecContext(ctx, `DELETE FROM recordings WHERE id = ?;`, id)
+		if err != nil {
+			return err
+		}
+	}
+
+	return tx.Commit()
+}
+
+// ListMergeableSegments returns recordings for a camera within a time window,
+// excluding merged and incomplete segments.
+func (d *DB) ListMergeableSegments(ctx context.Context, cameraID string, windowStart, windowEnd time.Time) ([]*model.Recording, error) {
+	rows, err := d.db.QueryContext(ctx,
+		`SELECT id, camera_id, file_path, format, started_at, ended_at, duration, file_size, frame_count, merged, archived FROM recordings WHERE camera_id = ? AND merged = 0 AND ended_at IS NOT NULL AND started_at >= ? AND started_at < ? ORDER BY started_at ASC;`,
+		cameraID, formatTime(windowStart), formatTime(windowEnd))
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var res []*model.Recording
+	for rows.Next() {
+		var r model.Recording
+		var startedAtStr, endedAtStr sql.NullString
+		if err := rows.Scan(&r.ID, &r.CameraID, &r.FilePath, &r.Format, &startedAtStr, &endedAtStr, &r.Duration, &r.FileSize, &r.FrameCount, &r.Merged, &r.Archived); err != nil {
+			return nil, err
+		}
+		r.StartedAt = scanTime(startedAtStr)
+		r.EndedAt = scanTime(endedAtStr)
+		res = append(res, &r)
+	}
+	return res, nil
+}
+
+// ListCameraMergeWindows returns hourly merge windows for a camera with 2+ segments.
+// Only includes recordings older than minAge.
+func (d *DB) ListCameraMergeWindows(ctx context.Context, cameraID string, minAge time.Duration) ([]MergeWindow, error) {
+	cutoff := time.Now().Add(-minAge).Format(sqliteTimeFormat)
+	query := `SELECT strftime('%Y-%m-%d %H', started_at) as hour, MIN(started_at), MAX(ended_at), COUNT(*), format FROM recordings WHERE camera_id = ? AND merged = 0 AND ended_at IS NOT NULL AND ended_at < ? GROUP BY hour, format HAVING COUNT(*) >= 2 ORDER BY hour ASC;`
+	rows, err := d.db.QueryContext(ctx, query, cameraID, cutoff)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var res []MergeWindow
+	for rows.Next() {
+		var w MergeWindow
+		var hourStr, minStart, maxEnd sql.NullString
+		if err := rows.Scan(&hourStr, &minStart, &maxEnd, &w.SegmentCount, &w.Format); err != nil {
+			return nil, err
+		}
+		w.StartTime = scanTime(minStart)
+		w.EndTime = scanTime(maxEnd)
+		res = append(res, w)
+	}
+	return res, nil
+}
+
+// UpsertCameraMerge writes per-camera merge config columns.
+// Pass nil pointers to leave fields unchanged (keep existing values).
+func (d *DB) UpsertCameraMerge(ctx context.Context, cameraID string, mergeEnabled *bool, mergeCheckInterval, mergeWindowSize, mergeMinSegmentAge *string, mergeBatchLimit, mergeMinSegmentsToMerge *int) error {
+	q := `UPDATE cameras SET
+		merge_enabled = COALESCE(?, merge_enabled),
+		merge_check_interval = COALESCE(?, merge_check_interval),
+		merge_window_size = COALESCE(?, merge_window_size),
+		merge_batch_limit = COALESCE(?, merge_batch_limit),
+		merge_min_segment_age = COALESCE(?, merge_min_segment_age),
+		merge_min_segments_to_merge = COALESCE(?, merge_min_segments_to_merge)
+		WHERE id = ?;`
+	_, err := d.db.ExecContext(ctx, q,
+		ptrToNullBool(mergeEnabled),
+		ptrToNullString(mergeCheckInterval),
+		ptrToNullString(mergeWindowSize),
+		ptrToNullInt64(mergeBatchLimit),
+		ptrToNullString(mergeMinSegmentAge),
+		ptrToNullInt64(mergeMinSegmentsToMerge),
+		cameraID)
+	return err
+}
